Replace `OnDomEvents.just()` with `stopDomEvents()` action
