+++ src/dom/actions/index.ts
@@ -1,2 +1,3 @@
 export * from './capture-dom-events';
 export * from './handle-dom-events';
+export * from './stop-dom-events';
